+++ src/features/binary-tree/components/RadiusButtons.tsx
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { ActionTypes, BinaryTreeContext } from "./BinaryTreeProvider";
+import { ActionTypes, BinaryTreeContext } from "../context/BinaryTreeProvider";
 import Button from "@mui/material/Button";
 import {
   Remove,
